requirements.txt
aiogram>=3.0.0
tortoise-orm>=0.20.0
aiosqlite>=0.19.0
aerich>=0.7.2
openai>=1.0.0
pydantic-settings>=2.0.0
tenacity>=8.2.0
apscheduler>=4.0.0a1
pytest>=7.4.0
pytest-asyncio>=0.23.0
